Split the navigation components
