+++ components/shared/shell/ProductNavigation.tsx
@@ -1,0 +1,10 @@
+import NavigationItems from './NavigationItems';
+import { NavigationProps, MenuItem } from './NavigationItems';
+
+const ProductNavigation = ({}: NavigationProps) => {
+  const menus: MenuItem[] = [];
+
+  return <NavigationItems menus={menus} />;
+};
+
+export default ProductNavigation;
